FIX: correcciones de la DB

+++ database/migrations/2021_03_09_160652_create_publications_table.php
@@ -17,8 +17,8 @@ class CreatePublicationsTable extends Migration
             $table->id();
             $table->string('title');
             $table->string('content');    
-            $table->integer('user_id')
-                ->unsigned(); 
+            $table->bigInteger('user_id')
+                ->unsigned();
             $table->foreign('user_id')
                 ->references('id')
                 ->on('users');
